+++ mex/editor/ingest/main.py
@@ -143,7 +143,7 @@ def search_input() -> rx.Component:
                 IngestState.go_to_first_page,
                 IngestState.refresh,
                 IngestState.resolve_identifiers,
-                IngestState.flag_imported_organizations,
+                IngestState.flag_ingested_items,
             ],
             style=rx.Style(margin="1em 0"),
             justify="center",
@@ -187,7 +187,7 @@ def search_results() -> rx.Component:
                 IngestState.results_transformed,
                 ingest_result,
             ),
-            pagination(IngestState, IngestState.flag_imported_organizations),
+            pagination(IngestState, IngestState.flag_ingested_items),
             spacing="4",
             custom_attrs={"data-testid": "search-results-section"},
             style=rx.Style(
@@ -270,7 +270,7 @@ def index() -> rx.Component:
                 IngestState.go_to_first_page,
                 IngestState.refresh,
                 IngestState.resolve_identifiers,
-                IngestState.flag_imported_organizations,
+                IngestState.flag_ingested_items,
             ],
             custom_attrs={"data-testid": "aux-tab-section"},
             style=rx.Style(