+++ tests/conftest.py
@@ -30,9 +30,9 @@ from mex.common.types import (
     Theme,
     YearMonthDay,
 )
+from mex.editor.api.main import api
 from mex.editor.settings import EditorSettings
 from mex.editor.types import EditorUserDatabase, EditorUserPassword
-from mex.mex import app
 
 pytest_plugins = ("mex.common.testing.plugin",)
 
@@ -65,7 +65,7 @@ def browser_type_launch_args(
 @pytest.fixture
 def client() -> TestClient:
     """Return a fastAPI test client initialized with our app."""
-    with TestClient(app.api, raise_server_exceptions=False) as test_client:
+    with TestClient(api, raise_server_exceptions=False) as test_client:
         return test_client