+++ CHANGELOG.md
@@ -46,7 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - warning when a page reflex navigates and there are still changes on the current page
 - added a switch to toggle all primary sources and values in the edit page
 - added textareas for fields listed in mex/editor/models.yaml
-- Warning when the create or edit browser window/tab is closed or navigated away
+- warning when the create or edit browser window/tab is closed or navigated away
 - new filter for reference field filtering
 - tab to switch between new reference field filter and the old had primary source filter